pyproject.toml
[project]
name = "video-subtitle-generator"
version = "0.1.0"
description = "一个基于 Streamlit 的视频字幕生成器，可以自动为视频生成中英文双语字幕"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "streamlit==1.32.2",
    "transformers==4.50.3",
    "torch==2.2.1",
    "openai==1.14.0",
    "python-ffmpeg==2.0.10",
    "moviepy==1.0.3",
    "srt==3.5.3",
    "python-dotenv==1.0.1",
]
requires-python = ">=3.11"
readme = "README.md"
license = "MIT"

[project.urls]
Homepage = "https://github.com/yourusername/video-subtitle-generator"
Repository = "https://github.com/yourusername/video-subtitle-generator.git"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/video_subtitle_generator"]

[tool.hatch.metadata]
allow-direct-references = true

[project.scripts]
video-subtitle-generator = "video_subtitle_generator.app:main"
